requirements.txt
# requirements.txt - Removed pgvector dependency
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.20
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
huggingface-hub==0.19.4
Pillow==10.1.0
numpy==1.24.3
asyncpg==0.29.0